refactor: LadderLinePoint domain.point로 패키지 이동
