+++ src/main/java/ladder/domain/point/LadderLinePoint.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.point;
 
 public class LadderLinePoint {
