+++ src/test/java/ladder/domain/point/LadderLinePointPainterTest.java
@@ -1,8 +1,5 @@
-package ladder;
+package ladder.domain.point;
 
-import ladder.domain.LadderLinePoint;
-import ladder.domain.LadderLinePointPainter;
-import ladder.domain.LadderLinePoints;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
