+++ src/test/java/ladder/domain/point/LadderLinePointsTest.java
@@ -1,7 +1,5 @@
-package ladder;
+package ladder.domain.point;
 
-import ladder.domain.LadderLinePoint;
-import ladder.domain.LadderLinePoints;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
